Start CDO BFO shapes from base branch
Note: A pre-2.0 version of BFO is reviewed due to properties having been
dropped for 2.0 with a plan to restore in the yet-unrelease 2.1.

Shapes have been adapted from the UCO BFO Profile repository, including
a few corrections found (copy-paste errors) while generating exemplar
data.

Some of the exemplars are known to be incomplete in their illustration,
because the demonstrated properties rely on a subproperty hierarchy.
Unfortunately, an error arose when attempting RDFS and OWL expansion
with `pyshacl`'s `--inference` flag, so an alternative expansion
mechanism will need to be devised to recognize that those demonstrations
are incomplete.

A follow-on patch will generate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>